requirements.txt
streamlit
streamlit-audiorecorder
numpy
scipy
plotly
Pillow
